ISSUE #3882 - destructure formModal
