+++ frontend/src/v5/ui/controls/headedModal/modalBody/modalBody.component.tsx
@@ -1,5 +1,5 @@
 /**
- *  Copyright (C) 2022 3D Repo Ltd
+ *  Copyright (C) 2023 3D Repo Ltd
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU Affero General Public License as
@@ -14,13 +14,21 @@
  *  You should have received a copy of the GNU Affero General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
+import { ScrollArea } from '@controls/scrollArea';
+import { FormDialogContent } from './modalBody.styles';
 
-import styled from 'styled-components';
-import { FormModal as FormModalBase } from '@controls/headedModal/formModal/formModal.component';
-import { FormDialogContent } from '@controls/headedModal/formModal/formModal.styles';
+export interface IModalBody {
+	children: any,
+	hideHorizontalScroll?: boolean,
+}
 
-export const FormModal = styled(FormModalBase)`
-	${FormDialogContent} {
-		padding: 7px 24px 22px;
-	}
-`;
+export const ModalBody = ({
+	children,
+	hideHorizontalScroll = true,
+}: IModalBody) => (
+	<ScrollArea variant="base" autoHeightMax="70vh" autoHeight hideHorizontal={hideHorizontalScroll}>
+		<FormDialogContent>
+			{children}
+		</FormDialogContent>
+	</ScrollArea>
+);
