+++ frontend/src/v5/ui/routes/dashboard/projects/federations/createFederationForm/createFederationForm.component.tsx
@@ -19,7 +19,7 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { formatMessage } from '@/v5/services/intl';
 import { NewFederationSettingsSchema } from '@/v5/validation/containerAndFederationSchemes/federationSchemes';
-import { FormModal } from '@controls/modal/formModal/formDialog.component';
+import { FormModal } from '@controls/headedModal/formModal/formModal.component';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { FormProvider, useForm } from 'react-hook-form';
 import { NewFederation } from '@/v5/store/federations/federations.types';
